api/python/requirements.txt
flask==3.0.2
flask-cors==4.0.0
openai==0.28.1
python-dotenv==1.0.1 